# shared/api/openapi.yaml
openapi: 3.0.1
info:
  title: Flex Work
  description: 'This is the REST API Documentation for ICBC Flex Work, by Team Flex'
  version: 0.0.10
servers:
  - url: 'http://34.224.250.122:8080/'
...
paths:
...
  /admin/upload-floor-data:
    post:
      summary:
        Upload a CSV of a floor's workspaces
      description:
        Upload a CSV containing information about each workspace on a floor (owner and features)
      tags:
        - Admin
      operationId: adminUploadFloorData
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                floorId:
                  type: integer
                floorData:
                  type: string
                  format: binary
      responses:
        '200':
          description: Successful Operation
        '400':
          description: Something went wrong
          content:
            text/plain:
              schema:
                type: string
              example:
                "Error: CSV was missing column: OfficeID"
  /admin/upload-floorplan-image:
    post:
      summary:
        Upload a JPG floor plan for a specific floor
      tags:
        - Admin
      operationId: adminUploadFloorplanImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                floorId:
                  type: integer
                floorplanImage:
                  type: string
                  format: binary
      responses:
        '200':
          description: Successful Operation
        '400':
          description: Something went wrong
          content:
            text/plain:
              schema:
                type: string
              example:
                "Error: Floor with ID 9000 does not exist"
...
  '/availabilities/{id}':
    delete:
      tags:
        - Office Lending
      summary: Cancel an Availability
      operationId: cancelAvailability
      parameters:
        - name: id
          in: path
          description: ID of the Availability to delete
          required: true
          explode: true
          schema:
            type: integer
          example: 23
      responses:
        '200':
          description: Successful Operation
          content: {}
        '400':
          description: Bad Request
          content: {}
...